+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.1.0 : 11.08.2025
 
+- **Feature**: Added transformations and hash utils.
 - **Feature**: Added main script, config and data files.
 - **Feature**: Initial project setup.